Add boost multindex implementation from joaquintides

+++ src/benchmarks/Insert.cpp
@@ -9,7 +9,7 @@ BENCHMARK(InsertHugeInt) {
         bench.beginMeasure("insert 100M int");
         Map<int, int> map;
         for (size_t n = 0; n < 100'000'000; ++n) {
-            map[rng()];
+            map[static_cast<int>(rng())];
         }
         bench.endMeasure(98841586, map.size());
 
@@ -21,14 +21,14 @@ BENCHMARK(InsertHugeInt) {
         auto const state = rng.state();
         bench.beginMeasure("reinsert 100M int");
         for (size_t n = 0; n < 100'000'000; ++n) {
-            map[rng()];
+            map[static_cast<int>(rng())];
         }
         bench.endMeasure(98843646, map.size());
 
         rng.state(state);
         bench.beginMeasure("remove 100M int");
         for (size_t n = 0; n < 100'000'000; ++n) {
-            map.erase(rng());
+            map.erase(static_cast<int>(rng()));
         }
         bench.endMeasure(0, map.size());
